% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seq}
\alias{seq}
\title{The input data are taken from the article by Sugauchi et al.(2002).
This paper describes recombination event observed in hepatitis B virus isolates.}
\format{
An object of S4 class. The multiple alignment of the 3 DNA sequences.
}
\source{
\href{https://jvi.asm.org/content/76/12/5985}{Journal of Virology}
}
\usage{
seq
}
\description{
The input data are taken from the article by Sugauchi et al.(2002).
This paper describes recombination event observed in hepatitis B virus isolates.
}
\examples{
data(hcv_C_Bj_Ba_sequences)
seq
}
\references{
Hepatitis B Virus of Genotype B with or without Recombination with Genotype C over the
Precore Region plus the Core Gene. Fuminaka Sugauchi et al. JOURNAL OF VIROLOGY, June 2002, p. 5985–5992.
10.1128/JVI.76.12.5985-5992.2002
}
\keyword{dataset}
